fix(app): move App to separate directory
